move splines to contrib
