+++ liesel/contrib/__init__.py
@@ -1,0 +1,7 @@
+"""
+Additional functionality that does not belong to the core of the Liesel framework.
+
+While we consider functionality in this module to be reliably functional, we may choose
+to organize it differently in the future. For example, submodules in ``liesel.contrib``
+might be moved into their own, separate libraries at some point in the future.
+"""
